Adds selective redux persistence
This also rearranges a bunch of stuff related to auth and loading user
data to better facilitate a good loading experience for the user.

When the app loads up and the user's blockstack creds are in local
storage, the app will show a "Loading user data..." message, followed by
a "Loading app data...", then the app will render with all of their user
data(from a blockstack) and all of their app data (previously saved via
the redux-persist library) immediately usable.

As before, the app will redirect the `/login` page if the user isn't
logged in.